Update SmartCar.ino
This update makes the smart car successfully do 5 laps to the right (clockwise) around an object. It added the gyroscope and reworked the entire movement of the car, making adjustments happen in real time continuously. All that is left to add is this final left loop and the turn around.